Refactor Ingress-specific lister code to separate file
